updates by Skala
merged:

  party-overlay: 1df85806e2432ff502ed871719bdc8cda73f9c94
  foo-overlay: cde51767054cbf2bb4bdaab8d0b04d8ea882bc6f
  bar-overlay: 09dbf39ca3130b8fa3d3c049518b56edd8dffcbe
  mate: d0a9dd0e6e833d37a7c69934c2b190fe75283a78
  squeezebox: 3b7b8b97ca62d555c9e8456eb685627976cab6c2
  funtoo-overlay: ea71ebe500b68e815f858a54529f0779ac183d16
  sabayon-for-gentoo: a7dd20051355af4d349f06658a12231343fda786

+++ eclass/mate.eclass
@@ -237,9 +237,14 @@ mate_src_install() {
 
 	unset MATECONF_DISABLE_MAKEFILE_SCHEMA_INSTALL
 
-	# Manual document installation
-	if [[ -n "${DOCS}" ]]; then
-		dodoc ${DOCS} || die "dodoc failed"
+	# Handle documentation as 'default' for eapi5 and newer, bug #373131
+	if has ${EAPI:-0} 0 1 2 3 4; then
+		# Manual document installation
+		if [[ -n "${DOCS}" ]]; then
+			dodoc ${DOCS} || die "dodoc failed"
+		fi
+	else
+		einstalldocs
 	fi
 
 	# Do not keep /var/lib/scrollkeeper because:
@@ -254,12 +259,20 @@ mate_src_install() {
 	rm -fr "${ED}/usr/share/applications/mimeinfo.cache"
 
 	# Delete all .la files
-	if [[ "${MATE_LA_PUNT}" != "no" ]]; then
-		ebegin "Removing .la files"
-		if ! { has static-libs ${IUSE//+} && use static-libs; }; then
-			find "${D}" -name '*.la' -exec rm -f {} + || die "la file removal failed"
+	if has ${EAPI:-0} 0 1 2 3 4; then
+		if [[ "${MATE_LA_PUNT}" != "no" ]]; then
+			ebegin "Removing .la files"
+			if ! { has static-libs ${IUSE//+} && use static-libs; }; then
+				find "${D}" -name '*.la' -exec rm -f {} + || die "la file removal failed"
+			fi
+			eend
 		fi
-		eend
+	else
+		case "${MATE_LA_PUNT}" in
+			yes)	prune_libtool_files --modules;;
+			no)		;;
+			*)		prune_libtool_files;;
+		esac
 	fi
 }
